+++ Libplanet/Action/Sys/Registry.cs
@@ -9,7 +9,7 @@ namespace Libplanet.Action.Sys
         {
             Mint = 0,
             Transfer = 1,
-            SetValidator = 2,
+            Initialize = 2,
         }
 
         public static IAction Deserialize(Bencodex.Types.Dictionary serialized)
@@ -48,7 +48,7 @@ namespace Libplanet.Action.Sys
             {
                 TypeId.Mint => new Mint(),
                 TypeId.Transfer => new Transfer(),
-                TypeId.SetValidator => new SetValidator(),
+                TypeId.Initialize => new Initialize(),
                 _ => throw new ArgumentOutOfRangeException(
                     nameof(typeId),
                     typeId,
@@ -61,7 +61,7 @@ namespace Libplanet.Action.Sys
             {
                 Mint _ => TypeId.Mint,
                 Transfer _ => TypeId.Transfer,
-                SetValidator _ => TypeId.SetValidator,
+                Initialize _ => TypeId.Initialize,
                 _ => throw new ArgumentException("Unknown system action type.", nameof(action)),
             };
     }